requirements.txt
Django==1.7.5
MySQL-python==1.2.5
PyMySQL==0.6.4
Werkzeug==0.10.1
django-extensions==1.5.0
djangorestframework==3.0.5
six==1.9.0
wsgiref==0.1.2
